fix: update utils method
